[EI-165] [api] rename: 파일명 오타 수정
